// src/generated/models/batch-patient-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FailedRegistration } from './failed-registration';
// May contain unused imports in some cases
// @ts-ignore
import type { PatientResponse } from './patient-response';

export interface BatchPatientResponse {
    /**
     * Total number of patients in the request
     */
    'totalRequested'?: number;
    /**
     * Number of successfully registered patients
     */
    'successCount'?: number;
    /**
     * Number of failed registrations
     */
    'failureCount'?: number;
    /**
     * List of successfully registered patients
     */
    'successfulRegistrations'?: Array<PatientResponse>;
    /**
     * List of failed registrations with error details
     */
    'failedRegistrations'?: Array<FailedRegistration>;
}